folder endpoint

+++ taskToDo-Backend/src/main/java/com/example/taskToDo/services/UserService.java
@@ -27,26 +27,35 @@ public class UserService {
   }
 
   // get User by Id
-  public Optional<User> getUserById(Long id){
-    return userRepository.findById(id);
+  public User getUserById(Long id){
+    Optional<User> user = userRepository.findById(id);
+    if(user.isPresent()){
+      return user.get();
+    }
+    return null;
   }
 
   // delete User by Id
-  public Optional<User> deleteUserById(Long id){
+  public User deleteUserById(Long id){
     Optional<User> userToDelete = userRepository.findById(id);
-    userRepository.deleteById(id);
-    return userToDelete;
+    if(userToDelete.isPresent()){
+      userRepository.deleteById(id);
+      return userToDelete.get();
+    }
+    return null;
   }
   
   // modify User by Id
-  public Optional<User> modifyUserById(Long id, User user){
+  public User modifyUserById(Long id, User user){
     Optional<User> userToModify = userRepository.findById(id);
     if(userToModify.isPresent()){
-      userToModify.get().setName(user.getName());
-      userToModify.get().setPassword(user.getPassword());
-      return userToModify;
+      User userModify = userToModify.get();
+      userModify.setName(user.getName());
+      userModify.setPassword(user.getPassword());
+      userRepository.save(userModify);
+      return userModify;
     } else {
-      return userToModify;
+      return null;
     }
   }
